chore: Ignore .tasks folder in prettier
Add .tasks/** to .prettierignore to prevent formatting checks on task files
