--- src/demos/dynamic-lesson.tsx
@@ -1,12 +1,0 @@
-export function DynamicLesson() {
-  return (
-    <div className="w-full h-full flex flex-col items-center justify-center p-8">
-      <h1 className="text-2xl font-bold">Dynamic Lesson Content</h1>
-      <p className="mt-4 text-muted-foreground">
-        This is a placeholder for dynamically generated lesson content.
-      </p>
-    </div>
-  )
-}
-
-export default DynamicLesson
