+++ src/server/payload/collections/Exercises/formatSlug.ts
