+++ src/ui/admin/ExerciseContentEditor/editors/InlineRichTextEditor.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React from 'react'
-import type { InlineRichText } from '@/shared/exercise-content/types'
+import type { InlineRichText } from '@/server/payload/collections/Exercises/types'
 import type { Media } from '@/payload-types'
 import {
   Bold,
