+++ src/ui/admin/ExerciseContentEditor/index.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import type { Media } from '@/payload-types'
-import { ExerciseBlockDefaults, generateId } from '@/shared/exercise-content/defaults'
-import type { ContentBlock } from '@/shared/exercise-content/types'
+import { ExerciseBlockDefaults, generateId } from '@/server/payload/collections/Exercises/defaults'
+import type { ContentBlock } from '@/server/payload/collections/Exercises/types'
 import { useField, useForm } from '@payloadcms/ui'
 import { Code, Copy, Image as ImageIcon, MoveDown, MoveUp, Plus, Trash2 } from 'lucide-react'
 import Image from 'next/image'
@@ -476,7 +476,9 @@ function renderQuestionEditor(
         canDelete={blockCount > 1}
       >
         <TrueFalseEditor
-          block={block as import('@/shared/exercise-content/types').QuestionSelectTrueFalseBlock}
+          block={
+            block as import('@/server/payload/collections/Exercises/types').QuestionSelectTrueFalseBlock
+          }
           onChange={onChange}
         />
       </QuestionBlockWrapper>
@@ -497,7 +499,9 @@ function renderQuestionEditor(
         canDelete={blockCount > 1}
       >
         <McqEditor
-          block={block as import('@/shared/exercise-content/types').QuestionSelectMcqBlock}
+          block={
+            block as import('@/server/payload/collections/Exercises/types').QuestionSelectMcqBlock
+          }
           onChange={onChange}
         />
       </QuestionBlockWrapper>
@@ -518,7 +522,9 @@ function renderQuestionEditor(
         canDelete={blockCount > 1}
       >
         <FreeResponseEditor
-          block={block as import('@/shared/exercise-content/types').QuestionFreeResponseBlock}
+          block={
+            block as import('@/server/payload/collections/Exercises/types').QuestionFreeResponseBlock
+          }
           onChange={onChange}
         />
       </QuestionBlockWrapper>
@@ -539,7 +545,7 @@ function renderQuestionEditor(
         canDelete={blockCount > 1}
       >
         <TableEditor
-          block={block as import('@/shared/exercise-content/types').QuestionTableBlock}
+          block={block as import('@/server/payload/collections/Exercises/types').QuestionTableBlock}
           onChange={onChange}
         />
       </QuestionBlockWrapper>
